@@ -9,11 +9,12 @@
     "axios": "^1.7.4",
     "bootstrap": "^5.3.7",
     "jwt-decode": "^4.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.3",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.1",
+    "react-dom": "^18.3.1",
+    "react-image-file-resizer": "^0.4.8",
     "react-infinite-scroll-component": "^6.1.0",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.28.0",
     "react-scripts": "^5.0.1",
     "react-textarea-autosize": "^8.5.3",
     "web-vitals": "^4.2.4"
@@ -49,6 +50,5 @@
   "engines": {
     "node": "18.x",
     "npm": "9.x"
-  },
-  "proxy": "http://localhost:8000/"
+  }
 }
