@@ -1,8 +1,8 @@
-import React, { useState } from "react";
-import { Link, useHistory } from "react-router-dom";
-import styles from "../../styles/SignInUpForm.module.css";
-import btnStyles from "../../styles/Button.module.css";
-import appStyles from "../../App.module.css";
+import React, { useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
+import styles from '../../styles/SignInUpForm.module.css';
+import btnStyles from '../../styles/Button.module.css';
+import appStyles from '../../App.module.css';
 
 import {
   Form,
@@ -12,21 +12,21 @@ import {
   Row,
   Container,
   Alert,
-} from "react-bootstrap";
-import axios from "axios";
-import { useRedirect } from "../../hooks/useRedirect";
+} from 'react-bootstrap';
+import axios from 'axios';
+import { useRedirect } from '../../hooks/useRedirect';
 
 const SignUpForm = () => {
-  useRedirect("loggedIn");
+  useRedirect('loggedOut');
   const [signUpData, setSignUpData] = useState({
-    username: "",
-    password1: "",
-    password2: "",
+    username: '',
+    password1: '',
+    password2: '',
   });
 
   const { username, password1, password2 } = signUpData;
   const [error, setError] = useState({});
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const handleChange = (e) => {
     setSignUpData({
@@ -38,8 +38,8 @@ const SignUpForm = () => {
   const handleSubmit = async (event) => {
     event.preventDefault();
     try {
-      await axios.post("/dj-rest-auth/registration/", signUpData);
-      history.push("/signin");
+      await axios.post('dj-rest-auth/registration/', signUpData);
+      navigate('/signin');
     } catch (err) {
       setError(err.response?.data);
     }
@@ -63,8 +63,8 @@ const SignUpForm = () => {
                 className='d-flex flex-column align-items-center'>
                 <Form.Label className='sr-only'>Username </Form.Label>
                 <Form.Control
-                  className={ styles.Input }
-                  placeholder="Username"
+                  className={styles.Input}
+                  placeholder='Username'
                   type='text'
                   name='username'
                   value={username}
@@ -83,7 +83,7 @@ const SignUpForm = () => {
                 className='d-flex flex-column align-items-center'>
                 <Form.Label className='sr-only'>Password</Form.Label>
                 <Form.Control
-                  className={ styles.Input }
+                  className={styles.Input}
                   placeholder='Password'
                   type='password'
                   name='password1'
@@ -101,37 +101,35 @@ const SignUpForm = () => {
               <Form.Group
                 controlId='password2'
                 className='d-flex flex-column align-items-center'>
-                <Form.Label className='sr-only'>
-                  Confirm Password
-                </Form.Label>
+                <Form.Label className='sr-only'>Confirm Password</Form.Label>
                 <Form.Control
-                  className={ styles.Input }
+                  className={styles.Input}
                   placeholder='Confirm Password'
                   type='password'
                   name='password2'
                   value={password2}
                   onChange={handleChange}
                 />
-              {error.password2?.map((message, index) => (
-                <Alert
-                variant='warning'
-                key={index}>
-                  {message}
-                </Alert>
-              ))}
-              <Button
-                className={`${btnStyles.Button} ${btnStyles.Green} ${btnStyles.FullWidth}`}
-                type='submit'>
-                Sign up
-              </Button>
-              {error.non_field_errors?.map((message, index) => (
-                <Alert
-                variant='warning'
-                className='mt-3'
-                key={index}>
-                  {message}
-                </Alert>
-              ))}
+                {error.password2?.map((message, index) => (
+                  <Alert
+                    variant='warning'
+                    key={index}>
+                    {message}
+                  </Alert>
+                ))}
+                <Button
+                  className={`${btnStyles.Button} ${btnStyles.Green} ${btnStyles.FullWidth}`}
+                  type='submit'>
+                  Sign up
+                </Button>
+                {error.non_field_errors?.map((message, index) => (
+                  <Alert
+                    variant='warning'
+                    className='mt-3'
+                    key={index}>
+                    {message}
+                  </Alert>
+                ))}
               </Form.Group>
             </Form>
           </Container>
